moved convert to ml into db
